requirements.txt
--extra-index-url https://pypi.fury.io/ericmichael/
omniagents[all]>=0.6.31
python-dotenv

# Platform integrations (install only the ones you need)
sendgrid>=6.11.0
twilio>=9.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-ads>=24.0.0

# Phase 4: Statistics and specialist LLM calls
scipy>=1.11.0
httpx>=0.27.0
